Move metadata completion to separate file
