+++ actions/fetch/handle-upload.js
@@ -5,8 +5,8 @@ import { Document } from 'yaml';
 import stylize from './stylize-doc.js';
 import apiToMetadata from './api-to-metadata.js';
 import Downloader from './downloader.js';
+import completeMetadata from './complete-metadata.js';
 import patchMetadata from './patch-metadata.js';
-import scrape from './scrape.js';
 import checkPreviousVersion from './check-previous-version.js';
 
 // # handleUpload(json)
@@ -116,39 +116,6 @@ export default async function handleUpload(json, opts = {}) {
 
 }
 
-// # completeMetadata(metadata, json)
-// Completes the metadata parsed from the api response with the description, 
-// images and subfolder. For the moment we need to use HTML scraping for this, 
-// as those fields are not yet included in the STEX api.
-async function completeMetadata(metadata, json) {
-
-	// Read the description, images & subfolder, and then complete the metadata 
-	// with it.
-	let { description, images, subfolder } = await scrape(json.fileURL);
-	let { package: pkg } = metadata;
-	let { info } = pkg;
-	if (!info.description) {
-		info.description = description;
-	}
-	if (!info.images) {
-		info.images = images;
-	}
-	pkg.subfolder = subfolder;
-
-	// Now generate the variants from what we've decided to include. This will 
-	// multiply the available variants by 2 in every step.
-	let includedVariants = findIncludedVariants(metadata);
-	let variants = generateVariants(includedVariants, metadata);
-
-	// If there are no variants, then we just include the assets as is.
-	let assets;
-	if (!variants || variants.length === 0) {
-		assets = metadata.assets.map(asset => ({ assetId: asset.assetId }));
-	}
-	Object.assign(metadata.package, { assets, variants });
-
-}
-
 // # serialize(json)
 function serialize(json) {
 	return json.map((json, index) => {
@@ -158,116 +125,4 @@ function serialize(json) {
 		}
 		return doc;
 	}).join('\n');
-}
-
-// # findIncludedVariants(metadata)
-// Finds the variants that are supported by this package. Note that we have no 
-// way yet of inspecting the actual package contents, we only support detecting 
-// this when there are separate downloads.
-function findIncludedVariants(metadata) {
-	let variants = [];
-	let { assets } = metadata;
-	if (assets.some(asset => /-(maxisnite|darknite)$/.test(asset.assetId))) {
-		variants.push('nightmode');
-	}
-	if (assets.some(asset => /-[rl]hd$/.test(asset.assetId))) {
-		variants.push('driveside');
-	}
-	if (assets.some(asset => /-cam$/.test(asset.assetId))) {
-		variants.push('CAM');
-	}
-	return variants;
-}
-
-// # generateVariants(variants, metadata)
-function generateVariants(variants, metadata) {
-	let output = [{}];
-	for (let variant of variants) {
-		let queue = output;
-		output = [];
-		let objects = generateVariant(variant, metadata);
-		for (let variant of objects) {
-			for (let q of queue) {
-				output.push({
-					variant: { ...q.variant, ...variant.variant },
-					dependencies: mergeArray(q.dependencies, variant.dependencies),
-					assets: mergeArray(q.assets, variant.assets),
-				});
-			}
-		}
-	}
-	return output.length < 2 ? undefined : output;
-}
-
-// # mergeArray(a, b)
-function mergeArray(a, b) {
-	let merged = [...(a||[]), ...(b||[])];
-	if (merged.length === 0) return;
-	return structuredClone(merged);
-}
-
-// # generateVariant(type, metadata)
-// Generates the essential variant information for the given type. For example, 
-// for day-and-night, there are two possible variants. Subsequently we'll merge 
-// this with the other variants.
-function generateVariant(type, metadata) {
-	let { assets } = metadata;
-	if (type === 'nightmode') {
-		return [
-			{
-				variant: { nightmode: 'standard' },
-				assets: [
-					{ assetId: findAsset(assets, 'maxisnite') },
-				],
-			},
-			{
-				variant: { nightmode: 'dark' },
-				dependencies: ['simfox:day-and-nite-mod'],
-				assets: [
-					{ assetId: findAsset(assets, 'darknite') },
-				],
-			},
-		];
-	} else if (type === 'CAM') {
-		return [
-			{
-				variant: { CAM: 'no' },
-			},
-			{
-				variant: { CAM: 'yes' },
-				assets: [
-					{ assetId: findAsset(assets, 'cam') },
-				],
-			},
-		];
-	} else if (type === 'driveside') {
-		return [
-			{
-				variant: { driveside: 'right' },
-				assets: [
-					{ assetId: findAsset(assets, 'rhd') },
-				],
-			},
-			{
-				variant: { driveside: 'left' },
-				assets: [
-					{ assetId: findAsset(assets, 'lhd') },
-				],
-			},
-		];
-	} else {
-		return [];
-	}
-}
-
-const regexes = {
-	maxisnite: /-m(axis)?n(ite)?$/,
-	darknite: /-d(ark)?n(ite)?$/,
-	lhd: /-lhd$/,
-	rhd: /-rhd$/,
-	cam: /-cam$/,
-};
-function findAsset(assets, id) {
-	const regex = regexes[id];
-	return assets.find(asset => regex.test(asset.assetId)).assetId;
 }
